requirements.txt
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
flake8==3.8.4
gears-cli==1.2.0
hiredis==1.1.0
mccabe==0.6.1
packaging==21.3
phone-iso3166==0.3.6
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==3.0.9
python-dotenv==0.15.0
redis==4.3.1
redisbloom==0.4.0
redistimeseries==1.4.3
rmtest==0.7.0
six==1.15.0
wrapt==1.14.1
